Modified receiptSchema and GET testing
- item.js -
Changed quantity and price type from String to Number

- server.test.js -
Discarded seed/soil keys from mockData structure in favor of a simpler Schema structure
Modified GET tests to reflect the simpler Schema
Added an Avocado item in the before() function (see issue below)

Issue (#2) still not resolved.  Originally, it was thought that the
tags:[String] was the reason why only 5 out of 6 objects were created.  After
further experimentation, it now appears that the before() function is
willing to create 5 or 7 objects, but not 6.  When either 5 or 7 objects are
created, the test tags:[String] is added to the object.